api/requirements.txt
uvicorn==0.28.0
fastapi==0.110.0
httpx==0.27.0
pytest==8.1.1
python-dotenv==1.0.1
sqlalchemy==2.0.28
openai==1.13.3
